Refactor backend to use finegrained visitors
This also moves more code out of the headers to their associated impl files
to improve compilation times and code navigation.
