+++ gcc/rust/backend/rust-compile-base.cc
@@ -20,9 +20,10 @@
 #include "rust-abi.h"
 #include "rust-compile-item.h"
 #include "rust-compile-stmt.h"
+#include "rust-compile-expr.h"
 #include "rust-compile-fnparam.h"
 #include "rust-compile-var-decl.h"
-
+#include "rust-constexpr.h"
 #include "rust-diagnostics.h"
 #include "rust-expr.h"	// for AST::AttrInputLiteral
 #include "rust-macro.h" // for AST::MetaNameValueStr
